feat: payments process done

+++ backend/app/schemas/payment.py
@@ -1,12 +1,13 @@
 from datetime import datetime
 
+from decimal import Decimal
 from pydantic import BaseModel, Field
 
 from app.schemas.common import ORMBaseSchema
 
 
 class PaymentCreate(BaseModel):
-    amount_paid: float = Field(
+    amount_paid: Decimal = Field(
         gt=0
     )
     payment_method: str = Field(
@@ -24,7 +25,7 @@ class PaymentResponse(ORMBaseSchema):
     payment_id: int
     debt_id: int
     created_by: int
-    amount_paid: float
+    amount_paid: Decimal
     payment_method: str
     notes: str | None
     paid_at: datetime
@@ -33,5 +34,5 @@ class PaymentResponse(ORMBaseSchema):
 
 class PaymentSummary(ORMBaseSchema):
     payment_id: int
-    amount_paid: float
+    amount_paid: Decimal
     paid_at: datetime
